requirements.txt
openai==0.28
tqdm
python-dotenv
sqlite-utils
